requirements.txt
streamlit
python-dotenv
nomic
chromadb
azure-ai-inference
langchain
pymupdf  